+++ coloraide/contrast/wcag21.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 from . import ColorContrast
 from ..types import AnyColor
 from typing import Any
+from ..cat import WHITES
 
 class WCAG21Contrast(ColorContrast):
     """WCAG 2.1 contrast ratio."""
@@ -16,8 +17,8 @@ class WCAG21Contrast(ColorContrast):
     def contrast(self, color1: AnyColor, color2: AnyColor, **kwargs: Any) -> float:
         """Contrast."""
 
-        lum1 = max(0, color1.luminance())
-        lum2 = max(0, color2.luminance())
+        lum1 = max(0, color1.luminance(white=WHITES['2deg']['D65']))
+        lum2 = max(0, color2.luminance(white=WHITES['2deg']['D65']))
         if (lum1 > lum2):
             lum1, lum2 = lum2, lum1
         return (lum2 + 0.05) / (lum1 + 0.05)